- Added function to forms - Added methods to get user input   - fps   - resolution - Fixed random video selection - Fixed session authentication - changed fps from decimal to integer - added/fixed new views

+++ Start/models.py
@@ -41,7 +41,7 @@ class User(models.Model):
     age = models.IntegerField()
     pixel_width = models.IntegerField()
     pixel_height = models.IntegerField()
-    fps = models.DecimalField(max_digits=4, decimal_places=3)
+    fps = models.IntegerField()
 
     class Meta:
         db_table = 'user'
